requirements.txt
aiogram==3.0.0b7
apscheduler
python-dotenv
asyncmy
SQLAlchemy
